refactor(inference): move kv cache domain into module
